Move list format command runner to not be windows only
